update: split legacy - re-organise
